Added unique IDs for multiple clients using same VID/PID
Uniquely identifying each board is necessary for allowing
clients to reenumerate connected interfaces using the 'bi' command.

Unfortunately there seems to be no consistent way to identify interfaces
accross different libraries/OSs. As a workaround, the equality method
was added to interfaces for library specific methods and is used to
hand out 16-bit IDs. This method was chosen to reduce bandwidth and
also allow changing to a two-byte bus/address pair if the option becomes
available.

VID/PID enumerations are then cached accross multiple clients so
device IDs are consistent and can be locked against for synchronization
in the server.