Added basic DAPLink layer functionality
Expanded both the client and server files into directories since they
were growing significantly.

Instead of returning a list of interface numbers with useful
information, client.getConnectedBoards now returns a list of
DAPLinkConnection objects with the same information stored in
attributes.

The DAPLinkConnection objects can be initialized for full DAPLink access.
By default they lock the connection, but can be explicitely unlocked for
multi-process access, although currently it's up to the user to
coordinate the multiple processes.

Also fixed bug in the IfSelection where __init__ was called on
preexisting instances.