Added read/write server commands
Currently, all reads are defered until a flush command is sent.
This simplifies the DAPLink interface while still allowing sequential
commands to be queued. It does cost an extra command for single
operations.

For ease of use, setDeferredTransfer functionality can be implemented
on the client side, with flush commands being sent to control when
transfers occur.