Modified handling objects with multiple implementations
Each object now provides a static name and availablilty attribute
that can be used to select specific implementations.

Each __init__ file contains a comprehension that chooses the best
available implementation and exposes it as default_implementation.
Every implementation that is available is exposed in the IMPLEMENTATION
dictionary.

These implementation options were then exposed to the higher level
interfaces, and tests were added for implementation options based on
availability.